[klas3bis] in epochX eemumu manu move MemoryAccess.h to ..
